move other server related stuff
