+++ .gitignore
@@ -1,1 +1,1 @@
-public/index.*
+server/public/index.*
