+++ package.json
@@ -11,7 +11,7 @@
     "dev": "nodemon server",
     "fmt": "yarn run prettier && yarn run eslint --fix",
     "test": "yarn eslint",
-    "eslint": "eslint --ignore-pattern=public/* .",
+    "eslint": "eslint --ignore-pattern=server/public/* .",
     "prettier": "prettier --write '{*,client/*,server/*}.js'",
     "heroku-postbuild": "webpack"
   },
